remove the deprecated static datacenters.yaml support
