+++ hack/run-api.sh
@@ -32,7 +32,7 @@ make kubermatic-api
 
 API_EXTRA_ARGS=""
 if [ "$KUBERMATIC_EDITION" == "ee" ]; then
-  API_EXTRA_ARGS="-dynamic-datacenters -dynamic-presets"
+  API_EXTRA_ARGS="-dynamic-presets"
 fi
 
 if [ -n "${CONFIG_FILE:-}" ]; then
