--- pkg/ee/cmd/kubermatic-installer/shared.go
@@ -1,68 +1,0 @@
-//go:build ee
-
-/*
-                  Kubermatic Enterprise Read-Only License
-                         Version 1.0 ("KERO-1.0”)
-                     Copyright © 2020 Loodse GmbH
-
-   1.	You may only view, read and display for studying purposes the source
-      code of the software licensed under this license, and, to the extent
-      explicitly provided under this license, the binary code.
-   2.	Any use of the software which exceeds the foregoing right, including,
-      without limitation, its execution, compilation, copying, modification
-      and distribution, is expressly prohibited.
-   3.	THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND,
-      EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-      MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-      IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
-      CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-      TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-      SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-   END OF TERMS AND CONDITIONS
-*/
-
-package kubermaticinstaller
-
-import (
-	"fmt"
-
-	"github.com/sirupsen/logrus"
-	"github.com/urfave/cli"
-
-	"k8c.io/kubermatic/v2/pkg/controller/operator/common"
-)
-
-var (
-	targetNamespaceFlag = cli.StringFlag{
-		Name:  "namespace",
-		Usage: "Namespace to use for created resources",
-		Value: "kubermatic",
-	}
-	unpauseSeedsFlag = cli.BoolFlag{
-		Name:  "unpause-seeds",
-		Usage: fmt.Sprintf("Do not set the %s annotation on every generated Seed, to facilitate easier incremental seed cluster upgrades", common.SkipReconcilingAnnotation),
-	}
-)
-
-func handleErrors(logger *logrus.Logger, action cli.ActionFunc) cli.ActionFunc {
-	return func(ctx *cli.Context) error {
-		err := action(ctx)
-		if err != nil {
-			logger.Errorf("❌ Operation failed: %v.", err)
-			err = cli.NewExitError("", 1)
-		}
-
-		return err
-	}
-}
-
-func setupLogger(logger *logrus.Logger, action cli.ActionFunc) cli.ActionFunc {
-	return func(ctx *cli.Context) error {
-		if ctx.GlobalBool("verbose") {
-			logger.SetLevel(logrus.DebugLevel)
-		}
-
-		return action(ctx)
-	}
-}
