+++ pkg/ee/cmd/seed-controller-manager/wrappers.go
@@ -26,28 +26,12 @@ package seedcontrollermanager
 
 import (
 	"context"
-	"flag"
 
-	eeprovider "k8c.io/kubermatic/v2/pkg/ee/provider"
 	"k8c.io/kubermatic/v2/pkg/provider"
 
 	ctrlruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-var (
-	dynamicDatacenters = false
-	datacentersFile    = ""
-)
-
-func AddFlags(fs *flag.FlagSet) {
-	fs.BoolVar(&dynamicDatacenters, "dynamic-datacenters", false, "Whether to enable dynamic datacenters. Enabling this and defining the datcenters flag will enable the migration of the datacenters defined in datancenters.yaml to Seed custom resources.")
-	fs.StringVar(&datacentersFile, "datacenters", "", "The datacenters.yaml file path.")
-}
-
 func SeedGetterFactory(ctx context.Context, client ctrlruntimeclient.Reader, seedName string, namespace string) (provider.SeedGetter, error) {
-	if dynamicDatacenters {
-		return provider.SeedGetterFactory(ctx, client, seedName, namespace)
-	}
-
-	return eeprovider.SeedGetterFactory(ctx, client, datacentersFile, seedName)
+	return provider.SeedGetterFactory(ctx, client, seedName, namespace)
 }
